JSON Project
Problem: Design a simple database query system. The system must accept a CSV file (students.csv) as input as the initial database. After initial loading of the file, the records must be sorted by their index. The query will be given in SQL with a restricted command set for simplicity (see below for details). The resulting data set (result of the query) must be stored as a JSON file.